pyproject.toml
[project]
name = "livia"
version = "0.1.0"
description = "Multimodal AI agent for Slack with integrated tools"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai-agents>=0.2.0",
    "openai>=1.35.0",
    "mcp>=1.9,<2",
    "slack-bolt>=1.20.1",
    "slack-sdk>=3.29.0",
    "aiohttp>=3.8.0",
    "certifi>=2024.8.30",
    "python-dotenv>=0.19.0",
    "pathlib2>=2.3.0",
    "tiktoken>=0.7.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
livia = "livia:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
asyncio_mode = "auto"